** Updated .NET target to 4.5 to use newer features
Added check to TX function to throw exception if the data in the command list seems wrong

Changed public definition for Epsilon_Command_List to ReadOnlyDictionary
Updated TOD handler class to accept this type, other code doesn't need to be changed

Modified EpsilonCommandInfo to include an internal read only flag which triggers an exception on all sets when enabled by the SetReadOnly function
This function is called by the command context class after populating the command list
Added a clone function to the commandinfo object to use when it's desireable to get a copy of an object (like in the command parser)
Added getter for Epsilon_Command_Info to prevent set accesses.

Modified the serialData_input function to call the clone function before modifying the commandinfo object in case of a decode failure.
This method is ugly, and previously it would modify the object in the main command list, this caused any future attempt to read the GPS time info to actually send our internal command ID for a decode failure (which returns a NACK)