Added concept of ReadRequirements to the EpsilonCommandInfo This is a list of command indexes that must be valid before issuing a read command on this command
The device context class now maintains a common list of all command contexts that have been read and decoded.

The transmit class called by the UI now attempts to resolve these requirements recursively before issuing the requested command.
Do not under any circumstances allow recursive dependencies to exist or this will break :)

Added version as a requirement to status and the new functions that can't be used on all series
Added GPS and version as requirements to read status
Added phase and leap info to requirements for GPS init